Added Collision position, grav panel, collidion begin/end

+++ ECSRpg/Engine/Physics/PhysicsCollisionWorldData.cpp
@@ -5,7 +5,7 @@
 PhysicsCollisionWorldData::PhysicsCollisionWorldData(Entity InEntity, Vector3 InPosition, Vector3 InHalfLimits, ColliderType InType, bool isTrigger)
 {
     entity = InEntity;
-    type = InType;
+    type = (int8)InType;
     IsTrigger = isTrigger;
 
     BoundingBox = CollisionAABB(InPosition, InHalfLimits);
